#8361: add sharded sweeps
